Remove unless event
