+++ kernel/metadata/core/src/main/java/org/apache/shardingsphere/metadata/persist/converter/GlobalRuleNodeConverter.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.config.rule.global.converter;
+package org.apache.shardingsphere.metadata.persist.converter;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
@@ -54,17 +54,28 @@ public final class GlobalRuleNodeConverter {
     /**
      * Is active version path.
      *
-     * @param ruleName rule name
      * @param rulePath rule path
      * @return version
      */
-    public static boolean isActiveVersionPath(final String ruleName, final String rulePath) {
-        Pattern pattern = Pattern.compile(getRuleNameNode(ruleName) + "/active_version$", Pattern.CASE_INSENSITIVE);
+    public static boolean isActiveVersionPath(final String rulePath) {
+        Pattern pattern = Pattern.compile(getRuleNameNode() + "/([\\w\\-]+)/active_version$", Pattern.CASE_INSENSITIVE);
         Matcher matcher = pattern.matcher(rulePath);
         return matcher.find();
     }
     
-    private static String getRuleNameNode(final String ruleName) {
-        return String.join("/", "", ROOT_NODE, ruleName);
+    /**
+     * Get rule name
+     *
+     * @param rulePath rule path
+     * @return rule name
+     */
+    public static Optional<String> getRuleName(final String rulePath) {
+        Pattern pattern = Pattern.compile(getRuleNameNode() + "/([\\w\\-]+)/active_version$", Pattern.CASE_INSENSITIVE);
+        Matcher matcher = pattern.matcher(rulePath);
+        return matcher.find() ? Optional.of(matcher.group(1)) : Optional.empty();
+    }
+    
+    private static String getRuleNameNode() {
+        return String.join("/", "", ROOT_NODE);
     }
 }
