+++ mode/type/cluster/core/src/main/java/org/apache/shardingsphere/mode/manager/cluster/coordinator/registry/config/watcher/NewGlobalRuleChangedWatcher.java
@@ -18,12 +18,13 @@
 package org.apache.shardingsphere.mode.manager.cluster.coordinator.registry.config.watcher;
 
 import org.apache.shardingsphere.infra.rule.event.GovernanceEvent;
-import org.apache.shardingsphere.infra.util.spi.ShardingSphereServiceLoader;
+import org.apache.shardingsphere.metadata.persist.converter.GlobalRuleNodeConverter;
 import org.apache.shardingsphere.metadata.persist.node.GlobalNode;
 import org.apache.shardingsphere.mode.event.DataChangedEvent;
 import org.apache.shardingsphere.mode.event.DataChangedEvent.Type;
+import org.apache.shardingsphere.mode.event.config.global.AlterGlobalRuleConfigurationEvent;
+import org.apache.shardingsphere.mode.event.config.global.DeleteGlobalRuleConfigurationEvent;
 import org.apache.shardingsphere.mode.manager.cluster.coordinator.registry.NewGovernanceWatcher;
-import org.apache.shardingsphere.mode.spi.GlobalRuleConfigurationEventBuilder;
 
 import java.util.Arrays;
 import java.util.Collection;
@@ -35,8 +36,6 @@ import java.util.Optional;
  * Global rule changed watcher.
  */
 public final class NewGlobalRuleChangedWatcher implements NewGovernanceWatcher<GovernanceEvent> {
-    
-    private static final Collection<GlobalRuleConfigurationEventBuilder> EVENT_BUILDERS = ShardingSphereServiceLoader.getServiceInstances(GlobalRuleConfigurationEventBuilder.class);
     
     @Override
     public Collection<String> getWatchingKeys(final String databaseName) {
@@ -54,12 +53,15 @@ public final class NewGlobalRuleChangedWatcher implements NewGovernanceWatcher<G
     }
     
     private Optional<GovernanceEvent> createGlobalRuleEvent(final DataChangedEvent event) {
-        for (GlobalRuleConfigurationEventBuilder each : EVENT_BUILDERS) {
-            Optional<GovernanceEvent> result = each.build(event);
-            if (!result.isPresent()) {
-                continue;
+        if (GlobalRuleNodeConverter.isActiveVersionPath(event.getKey())) {
+            Optional<String> ruleName = GlobalRuleNodeConverter.getRuleName(event.getKey());
+            if (!ruleName.isPresent()) {
+                return Optional.empty();
             }
-            return result;
+            if (Type.ADDED == event.getType() || Type.UPDATED == event.getType()) {
+                return Optional.of(new AlterGlobalRuleConfigurationEvent(ruleName.get(), event.getKey(), event.getValue()));
+            }
+            return Optional.of(new DeleteGlobalRuleConfigurationEvent(ruleName.get()));
         }
         return Optional.empty();
     }
